+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-6602] Opts to put cache and add delete functionality
+
 ## [6.5.2] 2024-04-29
 
 ### Fixed